Week 4 Notes
Week 4 notes
This is a test class that shows the difference between an immutable priority queue and a mutable priority queue.  The point is to show immutable PQs are the better approach and will keep data integrity/proper ordering.

The bottom of the file is some binary heap calculations to show where each node should be place in the heap based on current index positions, and where left and right child should be placed.